requirements.txt
SQLAlchemy~=1.4.31
passlib~=1.7.4
pydantic~=1.9.0
fastapi~=0.73.0
fastapi-utils~=0.2.1
uvicorn~=0.17.5
python-dotenv~=0.19.2
mariadb~=1.0.9
Jinja2~=3.0.3
PyJWT~=2.3.0
starlette~=0.17.1
